requirements.txt
Django==3.2.7
djangorestframework==3.12.4
PyJWT==2.3.0
six==1.16.0
Pillow==9.0.0